Refactor DFSHelper.java for readability and efficiency
Simplified the code in the DFSHelper.java file by streamlining methods convertNeighboursToLongs and amIStuck. The refactor removes the need for intermediate collections and explicit loop by making use of stream operations. I find this a little more confusing but best way to get comfortable with streams is to use them.