feat: Created a ExploreNode class to create and track the statuses of each visited node in the explore phase (the constructor contains logic to ensure a node that has been visited does not get added as unvisited)
Created a ParentMap class to contain the node and neighbours of said node - all interaction and requesting of data is done through methods that ensure data integrity.

Tested the ExploreNode and ParentMap fully before implementing them in code.

Renamed some stuff to make it fit with my new naming convention.